separate applet and sensemaker config
