+++ dnas/sensemaker/zomes/coordinator/sensemaker/src/applet.rs
@@ -2,8 +2,7 @@ use std::collections::BTreeMap;
 
 use hdk::prelude::*;
 use sensemaker_integrity::{
-    AppletConfig, ConfigCulturalContext, ConfigMethod, ConfigResourceType, CulturalContext,
-    Dimension, EntryTypes, LinkTypes, Method, ResourceType,
+    AppletConfig, AppletConfigInput, CulturalContext, EntryTypes, LinkTypes, Method, ResourceType,
 };
 
 use crate::{
@@ -11,19 +10,10 @@ use crate::{
     utils::entry_from_record,
 };
 
-#[derive(Clone, Serialize, Deserialize, Debug)]
-pub struct AppletConfigInput {
-    pub name: String,
-    // pub ranges: Vec<Range>, // leaving out ranges since this is not an entry and is just part of the dimension
-    pub dimensions: Vec<Dimension>,
-    // the base_type field in ResourceType needs to be bridged call
-    pub resource_types: Vec<ConfigResourceType>,
-    pub methods: Vec<ConfigMethod>,
-    pub cultural_contexts: Vec<ConfigCulturalContext>,
-}
-
 #[hdk_extern]
 pub fn register_applet(applet_config_input: AppletConfigInput) -> ExternResult<AppletConfig> {
+    // check the format of the applet config
+    applet_config_input.clone().check_format()?;
     // check that it doesn't already exist
     let maybe_applet_config = check_if_applet_config_exists(applet_config_input.name.clone())?;
     if let Some(applet_config) = maybe_applet_config {
@@ -31,46 +21,7 @@ pub fn register_applet(applet_config_input: AppletConfigInput) -> ExternResult<A
         Ok(applet_config)
     } else {
         // applet config doesn't exist, create it
-        let mut dimensions: BTreeMap<String, EntryHash> = BTreeMap::new();
-        for dimension in applet_config_input.dimensions {
-            dimensions.insert(dimension.name.clone(), create_dimension(dimension)?);
-        }
-        let mut resource_types: BTreeMap<String, EntryHash> = BTreeMap::new();
-        for config_resource_type in applet_config_input.resource_types {
-            resource_types.insert(
-                config_resource_type.name.clone(),
-                create_resource_type(ResourceType::try_from(config_resource_type)?)?,
-            );
-        }
-        let mut methods: BTreeMap<String, EntryHash> = BTreeMap::new();
-        for config_method in applet_config_input.methods {
-            methods.insert(
-                config_method.name.clone(),
-                create_method(Method::try_from(config_method)?)?,
-            );
-        }
-        let mut cultural_contexts: BTreeMap<String, EntryHash> = BTreeMap::new();
-        for config_context in applet_config_input.cultural_contexts {
-            cultural_contexts.insert(
-                config_context.name.clone(),
-                create_cultural_context(CulturalContext::try_from(config_context)?)?,
-            );
-        }
-        let applet_config = AppletConfig {
-            name: applet_config_input.name,
-            dimensions,
-            resource_types,
-            methods,
-            cultural_contexts,
-        };
-        create_entry(&EntryTypes::AppletConfig(applet_config.clone()))?;
-        let applet_config_eh = hash_entry(&EntryTypes::AppletConfig(applet_config.clone()))?;
-        create_link(
-            applet_config_typed_path(applet_config.name.clone())?.path_entry_hash()?,
-            applet_config_eh.clone(),
-            LinkTypes::AppletConfig,
-            (),
-        )?;
+        let (applet_config, _) = create_entries_from_applet_config(applet_config_input.clone())?;
         Ok(applet_config)
     }
 }
@@ -101,4 +52,53 @@ pub fn check_if_applet_config_exists(applet_name: String) -> ExternResult<Option
 
 fn applet_config_typed_path(applet_name: String) -> ExternResult<TypedPath> {
     Ok(Path::from(format!("all_applets.{}", applet_name)).typed(LinkTypes::AppletName)?)
+}
+
+// create all entries specified in the config
+pub fn create_entries_from_applet_config(
+    config: AppletConfigInput,
+) -> ExternResult<(AppletConfig, EntryHash)> {
+    // dimensions
+    let mut dimensions: BTreeMap<String, EntryHash> = BTreeMap::new();
+    for dimension in config.dimensions {
+        dimensions.insert(dimension.name.clone(), create_dimension(dimension)?);
+    }
+    let mut resource_types: BTreeMap<String, EntryHash> = BTreeMap::new();
+    for config_resource_type in config.resource_types {
+        resource_types.insert(
+            config_resource_type.name.clone(),
+            create_resource_type(ResourceType::try_from(config_resource_type)?)?,
+        );
+    }
+    let mut methods: BTreeMap<String, EntryHash> = BTreeMap::new();
+    for config_method in config.methods {
+        methods.insert(
+            config_method.name.clone(),
+            create_method(Method::try_from(config_method)?)?,
+        );
+    }
+    let mut cultural_contexts: BTreeMap<String, EntryHash> = BTreeMap::new();
+    for config_context in config.cultural_contexts {
+        cultural_contexts.insert(
+            config_context.name.clone(),
+            create_cultural_context(CulturalContext::try_from(config_context)?)?,
+        );
+    }
+
+    let applet_config = AppletConfig {
+        name: config.name,
+        dimensions,
+        resource_types,
+        methods,
+        cultural_contexts,
+    };
+    create_entry(&EntryTypes::AppletConfig(applet_config.clone()))?;
+    let applet_config_eh = hash_entry(&EntryTypes::AppletConfig(applet_config.clone()))?;
+    create_link(
+        applet_config_typed_path(applet_config.name.clone())?.path_entry_hash()?,
+        applet_config_eh.clone(),
+        LinkTypes::AppletConfig,
+        (),
+    )?;
+    Ok((applet_config, applet_config_eh))
 }
