+++ dnas/sensemaker/zomes/integrity/sensemaker/src/cultural_context.rs
@@ -1,8 +1,8 @@
 use hdi::prelude::*;
 
 use crate::{
+    applet::{ConfigCulturalContext, ConfigThreshold},
     dimension::RangeValue,
-    properties::{ConfigCulturalContext, ConfigThreshold},
     ResourceType,
 };
